isci: controller state machine cleanup step2
scic_sds_controller_set_base_state_handlers() hides the fact that there is
only one set of controller handlers.  Moving this knowledge directly to the
routines that invoke controller methods makes the code easier to analyze, and
simplifies struct scic_sds_controller.

It also adds an additional read to find the handler method, need to trace if
this change increases the number of instructions required to submit / retire
an i/o.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>