isci: move controller allocation to the core (kill isci_module_struct)
Most of isci_module_struct is unused, and the bits that are used
(library handle and library memory) are not needed if we teach the core
how to allocate a controller.

The library interface also initializes the per-object logger masks, but
those initializations are not needed becuase we don't use the per-object
mask functionality of the logger and want to replace that with
tracepoints eventually.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>