isci: cleanup scic_remote_device_construct
While making the kmem_cache global noticed that isci_remote_device is
maintaining two pointers to the core object when only one is needed.
scic_remote_device_construct is giving us back the same pointer we
handed in after type casting so just eliminate that indirection and
parameter.

This also makes scic_remote_device_construct typesafe / removes a HANDLE_T
usage.

The memory passed to construct is assumed to be zeroed so we don't need to
init it twice.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>