isci: use a module scope kmem_cache rather then per-host
There is no need for per-host caches and a singleton allows us to delete
the cache_name field from isci_host.

This cleans up a memory leak in isci_module_init (transport template
release on pci_register_driver failure) and fixes the return value to be
a valid error code.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>