isci: remove 'library' dependency for initializing registers
Instead of telling the library the pci revision and core index only to have it
ask for the same information back, just teach the driver how to init the
register pointers directly.

1/ rename isci_pci_func/isci_pci to isci_pci_info/pci_info
2/ introduce a pdev to pci_info helper and delete isci_pci_info pointer in
   isci_host
3/ delete bar parsing code and instead pass pci_iomap values to
   scic_controller_construct and move register initialization from
   controller_initialize to construct.

Note: the comment around scic_sds_controller_reset_hardware()
	/*
	 * / todo: This should really be done in the reset state enter but
	 * /       the controller has not yet been initialized before getting
	 * /       to the reset enter state so the PCI BAR is not yet assigned */
is stale because we already call it from
scic_sds_controller_resetting_state_enter()... in any event we now
initialze the registers much earlier so it is not a concern.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>