ARM: DMA-API: only do one dsb per DMA operation
Currently, we bury the dsb in the depths of the CPU DMA implementation.
This has the side effect that we use this barrier multiple times when
mapping or unmapping SG lists which is not necessary.

As dsb can be expensive on some platforms, lets move this so we only
perform one dsb per DMA mapping/unmapping operation.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>