ARM: poison initmem when it is freed
When the initmem is freed, we can no longer rely on its contents.  In
lightly loaded systems, this memory may persist for some time, making
it harder discover run-time issues (caused by the build warnings being
ignored.)

Poison the initmem at the point where it is freed to encourage run-time
problems when initmem is dereferenced as an aid to finding such issues.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>