video: atmel_hlcdfb: now, suspend/resume is managed only by the base layer
When going to suspend or resume the functions will be called by the base layer
but also by the overlay layers. It's not necessary to do it several time.
Moreover the base address for overlays is not the base address of the IP so
offsets to access some registers are wrong.

Signed-off-by: Ludovic Desroches <ludovic.desroches@atmel.com>