Change the internal SRAM memory type MT_MEMORY_NONCACHED
Because MT_DEVICE is not executable in armv7, we change
the the internal SRAM memory type to MT_MEMORY_SO as commit
5fe012d4aee98ef82df66e6763882c735902ed0f,
but MT_MEMORY_SO is strongly ordered, this feature is not required
for the internal SRAM memory.

So we change the internal SRAM memory type to MT_MEMORY_NONCACHED.
As it seems that caching this internal SRAM memory is not necessary,
we chose the this memory type.

Signed-off-by: Wenyou Yang <wenyou.yang@atmel.com>