pinctrl: at91: don't disable pin interrupt when enabling muxing function
Before enabling a muxing function, pin interrupt is disabled (probably
to prevent spurious interrupts). It can causes several issues mainly for
power management when switching from default to sleep state or the other
way around. So if the irq was enabled for a pin, restore it once pin has
been configured.

Signed-off-by: Ludovic Desroches <ludovic.desroches@atmel.com>