media: atmel-isc: fix out-of-buffers freeze bug
Before this patch, when the buffer run out, there is no place to start_dma()
again. So that will cause the driver freezed.

This patch will check whether buffer is run out in buffer_queue(). if yes, then
start_dma() again.

buffer_queue() will be called when new buffer is available, so it's right place
to check buffer is run out or not.

Signed-off-by: apoggiali <antonio.poggiali@datalogic.com>
[refined the commit log]
Signed-off-by: Josh Wu <josh.wu@atmel.com>